Move "basic type tests" out of the overall "connection test" file
